Adapt Orbit chain e2e tests
