+++ scripts/testSetup.ts
@@ -1,53 +1,52 @@
 import { JsonRpcProvider } from '@ethersproject/providers'
-import { L1Network, L2Network, addCustomNetwork } from '@arbitrum/sdk'
+import { L1Network, L2Network } from '@arbitrum/sdk'
 import { execSync } from 'child_process'
 import { Bridge__factory } from '@arbitrum/sdk/dist/lib/abi/factories/Bridge__factory'
 import { RollupAdminLogic__factory } from '@arbitrum/sdk/dist/lib/abi/factories/RollupAdminLogic__factory'
 
-export const config = {
-  arbUrl: 'http://localhost:8547',
-  ethUrl: 'http://localhost:8545',
+export function sleep(ms: number) {
+  return new Promise(resolve => setTimeout(resolve, ms))
 }
 
-export const getCustomNetworks = async (
+export const getLocalNetworks = async (
   l1Url: string,
   l2Url: string
 ): Promise<{
   l1Network: L1Network
-  l2Network: Omit<L2Network, 'tokenBridge'> & { nativeToken: string }
+  l2Network: Omit<L2Network, 'tokenBridge'>
 }> => {
   const l1Provider = new JsonRpcProvider(l1Url)
   const l2Provider = new JsonRpcProvider(l2Url)
   let deploymentData: string
 
+  let data = {
+    bridge: '',
+    inbox: '',
+    'sequencer-inbox': '',
+    rollup: '',
+  }
+
   let sequencerContainer = execSync(
-    'docker ps --filter "name=sequencer" --format "{{.Names}}"'
+    'docker ps --filter "name=l3node" --format "{{.Names}}"'
   )
     .toString()
     .trim()
 
   deploymentData = execSync(
-    `docker exec ${sequencerContainer} cat /config/deployment.json`
+    `docker exec ${sequencerContainer} cat /config/l3deployment.json`
   ).toString()
 
-  const parsedDeploymentData = JSON.parse(deploymentData) as {
+  data = JSON.parse(deploymentData) as {
     bridge: string
     inbox: string
     ['sequencer-inbox']: string
     rollup: string
-    ['native-erc20-token']: string
   }
 
-  const rollup = RollupAdminLogic__factory.connect(
-    parsedDeploymentData.rollup,
-    l1Provider
-  )
+  const rollup = RollupAdminLogic__factory.connect(data.rollup, l1Provider)
   const confirmPeriodBlocks = await rollup.confirmPeriodBlocks()
 
-  const bridge = Bridge__factory.connect(
-    parsedDeploymentData.bridge,
-    l1Provider
-  )
+  const bridge = Bridge__factory.connect(data.bridge, l1Provider)
   const outboxAddr = await bridge.allowedOutboxList(0)
 
   const l1NetworkInfo = await l1Provider.getNetwork()
@@ -63,17 +62,16 @@ export const getCustomNetworks = async (
     isArbitrum: false,
   }
 
-  const l2Network: Omit<L2Network, 'tokenBridge'> & { nativeToken: string } = {
+  const l2Network: Omit<L2Network, 'tokenBridge'> = {
     chainID: l2NetworkInfo.chainId,
     confirmPeriodBlocks: confirmPeriodBlocks.toNumber(),
     ethBridge: {
-      bridge: parsedDeploymentData.bridge,
-      inbox: parsedDeploymentData.inbox,
+      bridge: data.bridge,
+      inbox: data.inbox,
       outbox: outboxAddr,
-      rollup: parsedDeploymentData.rollup,
-      sequencerInbox: parsedDeploymentData['sequencer-inbox'],
+      rollup: data.rollup,
+      sequencerInbox: data['sequencer-inbox'],
     },
-    nativeToken: parsedDeploymentData['native-erc20-token'],
     explorerUrl: '',
     isArbitrum: true,
     isCustom: true,
@@ -88,45 +86,4 @@ export const getCustomNetworks = async (
     l1Network,
     l2Network,
   }
-}
-
-export const setupNetworks = async (l1Url: string, l2Url: string) => {
-  const { l1Network, l2Network: coreL2Network } = await getCustomNetworks(
-    l1Url,
-    l2Url
-  )
-  const l2Network: L2Network & { nativeToken: string } = {
-    ...coreL2Network,
-    tokenBridge: {
-      l1CustomGateway: '',
-      l1ERC20Gateway: '',
-      l1GatewayRouter: '',
-      l1MultiCall: '',
-      l1ProxyAdmin: '',
-      l1Weth: '',
-      l1WethGateway: '',
-
-      l2CustomGateway: '',
-      l2ERC20Gateway: '',
-      l2GatewayRouter: '',
-      l2Multicall: '',
-      l2ProxyAdmin: '',
-      l2Weth: '',
-      l2WethGateway: '',
-    },
-  }
-
-  addCustomNetwork({
-    customL1Network: l1Network,
-    customL2Network: l2Network,
-  })
-
-  return {
-    l1Network,
-    l2Network,
-  }
-}
-
-export function sleep(ms: number) {
-  return new Promise(resolve => setTimeout(resolve, ms))
 }
